Split GPOS/GSUB builders, move to layout module
No functional change, just restructuring things.
