+++ fea-rs/src/compile.rs
@@ -4,7 +4,6 @@ use self::{compile_ctx::CompilationCtx, validate::ValidationCtx};
 
 pub use output::Compilation;
 
-mod builders;
 mod compile_ctx;
 mod consts;
 mod glyph_range;
